[AUTO BATCH PLUGIN] rename the src files of auto batch plugin
