+++ src/plugins/auto_batch/CMakeLists.txt
@@ -8,15 +8,15 @@ endif()
 
 set(TARGET_NAME "openvino_auto_batch_plugin")
 
-file(GLOB SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/auto_batch_*.cpp)
+file(GLOB SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
 
-file(GLOB HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/src/auto_batch_*.hpp)
+file(GLOB HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/src/*.hpp)
 
 ov_add_plugin(NAME ${TARGET_NAME}
               DEVICE_NAME "BATCH"
               PSEUDO_DEVICE
               SOURCES ${SOURCES} ${HEADERS}
-              VERSION_DEFINES_FOR src/auto_batch_plugin.cpp ADD_CLANG_FORMAT)
+              VERSION_DEFINES_FOR src/plugin.cpp ADD_CLANG_FORMAT)
 
 target_link_libraries(${TARGET_NAME} PRIVATE Threads::Threads)
