+++ src/plugins/auto_batch/src/plugin.hpp
